Add graphics algo to convert alpha mask texture to rgb texture
Add graphics algo to convert alpha mask texture to rgb texture. The
intended use case is to use this to convert the alpha mask to rgb
texture when the texture needs to be bound to an ES2 FBO as a render
target. This function converts the alpha mask to a rgb texture,
copies the alpha channel over and sets the logical alpha mask flag
on the result texture.