Refactor UI open/close action handling in the engine
Refactor UI open/close action handling in the engine. The handling
must now go through a queue in order to make sure that things
always happen in the intended order. For example if the game does
the following calls:

  Game:CloseUI(ui, 0)
  Game:OpenUI('Menu')

Menu cannot be opened and pushed on the UI stack until the previous
UI has completed closing and has finished running the UI animations.

For open this problem doesn't exist since we can simply always push
another UI onto the stack and when a previous UI is then resumed the
Open state can continue from where it was left off. So this means that
it's not a "must" to wait for the open to complete (in fact no such
mechnism exists right now) until the next UI can be opened.

Additional and related problem is the UI stack manipulation and
accidental superfluous requests. For example if the game is closing
UIs on Escape key press and calls CloseUI on each key press we'll
end up posting several close UI requests in the queue thus resulting
(possibly) closing UIs that were not intended to be closed. Unless
of course there's some filtering of UI input.

Since currently the UI cannot provide such transient state to the
game the only potential way to do this is to implement the filtering
in the engine by checking the current UI stack action queue for
duplicate requests and then rejecting the dupes. The correct
behaviour also requires the game to call CloseUI(ui, 0) instead of
CloseUI(0) to make sure that the filtering works as intended.