Add Python utility script in the game package output
Add Python utility script in the game package output. The script
isn't needed for deployment but only for testing locally in the
browser. The script will help with setting the right HTTP headers
in order to turn on the right policies in the user agent for best
performance and features.