Add function to test for script var "sameness" in value
Add function to test for script var "sameness" in value. Use the
sameness to clean scene entity instance variable overrides that
are actually the same value as the class value.

TODO: for floats this will probably fail and some kind of epsilon
check could be used to see if the floats are "close enough" to be
the same.