Tweak the entity instance dialog UI and layout
Tweak the entity instance dialog UI and layout. Show the entity
class scripting variable values if there are no entity instance
overrides for those values.

TODO: Do the same change on the flags etc. Show the class value
if no instance value is available and then use the "sameness" of
the value setting to discard overrides that are the same as the
class value.