Refactor the "known" set of material uniforms in the material class
Refactor the "known" set of material uniforms in the material class.
The implementation can be simplified by using the uniform map that
exists for the custom uniforms. The old API for accessing the built-in
shader uniforms can be kept in place and the implementation is just
refactored to store the uniforms in the map. This also saves some
space in the JSON since values that aren't used don't need to be
stored (the defaults are applied at uniform set time).