Add workspace material migration path for removing unused uniforms
Add workspace material migration path for removing unused uniforms.
The material no longer keeps all the uniform data around for the
set of "known" uniforms. Instead the uniforms are only stored when
the value has been changed by the user.

This means that when loading data from previous data files if the
uniform values that are loaded are the same (or close enough) as the
current defaults the value can be nuked from the set of uniforms.