Refactor shader source and object generation and compilation
Refactor shader source and object generation and compilation. This
commit refactors the way the drawables and materials provide shaders
from providing gfx::Shader* objects to shader source. The compilation
is then moved from drawables and materials into the painter.

Additionally the "shader pass" abstraction is removed and refactored
into a shader program that is used by the painter to create the
final shader sources from the sources coming from the materials and
drawables.

The key motivations behind this complicated change are:

- Simplifies the materials and drawables when they only have to
  concern themselves with transforming their state + incoming state
  into a shader source specific to that state.

- Refactors and simplifies the need to be able to generate various
  types of shaders for different types of render passes. Using the
  render pass abstraction inside the drawables/materials is a wrong
  abstraction. A better solution is to have a higher level ShaderProgram
  abstraction that can provide the final source either by completely
  replacing the material/drawable source or combining them together
  in some meaningful way.

Together these changes should allow for better way to forward to
create more complicated render passes such as depth pre-pass or
light pass.

For example a shader program that performs lighting could define an
GLSL interface for taking on the material properties. The material
will then generate shader compatible with that GLSL interface and the
shader program will then fill in the light computation part using
the material properties. This design should allow these two to vary
independently.

TODO:

- Maybe the dynamic/static material and drawable state setting should
  go through the ShaderProgram object then?

- Add some helper information about the high level GLSL interface
  requirement when generating the drawable / material shader source.

- Remove the GLSL main functions from the drawable and material
  shader sources. Let the ShaderProgram provide the GLSL "main"
  and then call the drawable / material "main" functions.