Add a setting for separating dev time game home from actual home
Add a setting for separating dev time game home from actual home
and add an alternative play action for cleaning the game home
before running the game. This feature is convenient for testing
features such as game state loading etc that require a "clean"
state to begin from the start.