Add a missing check for mipmap generation with sRGB textures on WebGL2
Add a missing check for mipmap generation with sRGB textures on WebGL2.
It would have seemd that this would be in fact supported on WebGL2
but when testing both chromium and firefox seem to have a problem.
chromium prints a GL_INVALID_OPERATION as well.

I think we need to do a workaround here but assuming the workdaround
is going to require texture read back and mipmap geneeration on the
CPU the operation must be combined with a resource pre-load step
that doesn't yet exist in the engine.