Add a workaround for random runtime glib assertion failures
Add a workaround for rnadom runtime glib assertion failures that
started happening after a system upgrade (likely because of GCC
changed and/or the libraries involved were updated). Seems someone
took a dookie all over the global state and randomly shit begins
to fail because initialization order is now different than what
someone expected.

The workaround/fix is to shuffle the libraries around when passing
them to the linker. What a fucking shit festival!