Refactor tilemap drawing and packet generation in renderer
Refactor tilemap drawing and packet generation in renderer so that
tilemap based draw packet generation is only in one place.

Refactor tilemap drawing to support editor also work using the
draw packets instead of tile batches. This currently needs to be
a separate draw function because of the complications of indicating
selection through edge detection in the editor.