Add first version of non-axis aligned drawing in renderer
Add first version of non-axis aligned drawing in renderer for
draw packets generated from entity drawables. The drawable item
associated with some entity node can now provide the renderer
with a value that defines whether the rendering view on the shape
should be axis aligned or dimetric and whether the projection
should be orthographic or perspective.

This change does not (or should not) change the logical game world
that applies to all other game logic (including the scripts and
script interfaces) and other sub-systems such as physics and spatial
partitioning. These systems always operate in 2D space, only rendering
can operate in 3D.