Add debug drawable for showing auxiliary drawable geometry data
Add debug drawable for showing auxiliary drawable geometry data.
The debug drawable wraps an actuald drawble and uses the buffer
geometry to capture the generated geometry data, vertices, indices
and draw commands and then applies some algorithm on the data in
order to produce some alternative/additional renderable geometry
related to the original drawable. The intended use cases are for
example showing a wireframe mesh that visualizes the triangles
or a normal mesh that visualizes each vertex normal.