Refactor the "endless" isometric grid rendering in the editor
Refactor the "endless" isometric grid rendering in the editor. The
problem is that using isometric painting and model transformation
we can run into near/far plan clipping issues when the grid is offset
around tile plane position that maps too far from the camera.

The solution for having the grid always cover the whole viewport
has been to take the center point of the window and map it to a
position on the tile plane. When the grid geometry is rendered
around this point it always covers the whole viewport. But When the
camera moves up the center point of the window (the projection surface)
maps to a tile map coordinate that is further away from the camera.
Using this point as the anchor point for the grid rendering will
then hit clipping issues at some distance.

This commit refactors the isometric grid case so that instead of
drawing the grid on the tile plane with the tile painter we project
the anchor coordinate (in tile coordinates) back to the scene plane
coordinates (remember the scene plane is parallel to the projection
plane and has no near/far plane clipping issues) and then apply
a different transformation that produces the same outcome as a
"real" isometric (dimetric) transformation + projection would.