Reimplement draw packet level culling in the renderer
Reimplement draw packet level culling in the renderer. The culling
is actually a bit more complicated than we might think:

- We must consider the drawable before deciding to cull or not. The
  node and its box aren't enough except in a limited number of cases.
  I.e. when the drawable is orthographic and axis aligned 2D shape.

- Checking against Camera::viewport must be done in a consistent
  coordinate space, i.e. the viewport is defined in the *game*
  coordinate space. Transforming the drawables by the view transform
  (that includes a rotation) requires the viewport to be rotated then
  too.

- View rotation needs to be considered before culling Rotating the
  camera is the same as rotating the world the other way.

Currently the simplest solution that catches all cases, all shapes,
all projections and viewing angles is to cull against the NDC by
taking the shape's bounding box's vertices and transforming them
into NDC and then applying  a simple check against the X and Y
bounds (-1.0f - 1.0f ). This should work since we know that even
with perspective projection the perspective correction is applied
so that the final projection is still orthographic.