Add a geometry draw command class for sub-mesh rendering
Add a geometry draw command class for sub-mesh rendering. The
geometry has all the draw commands but those are defined so that
all the elements in the geometry are drawn. The draw command class
encapsulates the geometry and provides a wrapper that overrides the
draw commands by defining an optional range of geometry draw commands.

This feature will allow sub-mesh rendering to take place so that
the whole model geometry is in a single geometry and particular
sub-mesh (typically requires it's own material) will/can then be
submitted as a separate draw call with the appropriate state,
program and draw command that defines only the draw commands that
are related to that geometry sub-mesh.