Refactor polygon mesh class and remove polygon building API
Refactor polygon mesh class and remove polygon building API to a
separate polygon builder class. Refactor the inline data inside
the polygon to be an optional chunk of data. The future will bring
a way to associate a data file URI with the polygon for larger
polygon meshes with more data, but for the time being the inline
data is convenient for loading small model (shapes).