Add gfx algo to extract color values from source texture
Add gfx algo to extract color values from source texture based on
the magnitude of the color value against target color value.

This functionality can be used to implement bloom effect by taking
the rendered output and extracting the highlights into a separate
texture which then can be blurred and blended on top of the scene.