Add missing renderer names to editor main widgets
Add missing renderer names to editor main widgets. The renderer
name is used as a prefix to identify GPU resources. Without the name
it's possible that each widget will map to the same GPU resource
which might cause unwanted contention. Or alternatively be a
blessing by reducing memory usage. Not sure yet...