Add a warmup rendering loop for GPU resource pre-loading
Add a warmup rendering loop for GPU resource pre-loading. Currently
there isn't a good mechanism to do straighforward pre-loading/loading
step for getting the GPU assets loaded. The material and drawable
system needs to be enhanced with better packing in order to create
some kind of pre-load list of data needs to be uploaded to the GPU.

This commit implements a little workaround by iterating over entities
and their nodes with drawable items and rendering each drawable
item with the assigned geometry and material a few times in order
to precipiate the GPU resource creation.