Refactor how performance tracing is enabled / disabled
Refactor how performance tracing is enabled / disabled. Control
the tracing in the base/trace with a separate thread local boolean
flag so that we can have both

  a) tracer object enabled and ready to go
  b) control whether the actual tracing takes place or not

Add a new Lua API so that the game can control the tracing
programmatically. This should be useful in cases where we can already
see that there's some performance issue for example  when spawning
some entity etc or performing some action so the tracing can be
started conditionally on some event and stopped later on in order
to try to reduce the volume of data recorded and make it easier to
see the interesting part of runtime behaviour in the trace.

Because the tracing functionality can also be controlled by the
debug/developer UI (when this is enabled) the tracing is implemented
via counter in order to have proper nesting of calls. Tracing is
only turned off when the counter reaches zero, i.e. when the number
of "disable" requests matches previous "disable" requests. Since
the requests are async (queued from the game and the UI) this design
should make sure that the world is in a consistent state.

Some problems related to the current tracing implementation:

- The tracing state must be managed carefully and the tracing must
  be started / stopped outside of any calls to trace so that the
  nesting of the trace stack works correctly. This in turn means that
  the any request to turn the tracing on/off must be delayed until
  the next frame start and before any tracing scopes are entered.

  This means we have one frame of latency which needs to be kept in
  the mind when trying to capture stuff.