Restructure the native main window setup wrt engine setup
Restructure the native main window setup wrt engine setup. First
setup the window based, i.e. size and positio and then setup the
engine. This avoids having to notify the engine about changes.

Move the engine window listener connection later in order to avoid
causing weird issues with window events happening before the engine
has been initialized.

Add some desperate checks in the functions to center the window and
to restore to previous size and position to check whether the window
would end up larger than the screen or at a position that is off
screen and then try to adjust to more reasonable values.

Also it seems that WDK has a Win32 bug, if the call to window::SetSize
specifies width/height that match the previous surface width/height
we have a weird issue that the window disappears and is never
properly shown. The window appears in the taskbar and receives
keyboard events etc. but just cannot been seen! :o