Refactor GPU program state away from the program object
Refactor GPU program state away from the program object. This change
enables the following features

- Possible to cache/serialize the program state and re-use the same
  later without having to reset all the uniform and texture state.

- Step towards making the program client API immutable.
  -> Immutable resources will let us capture the whole draw state
     completely
  -> Allows the draw calls to be re-organized and resource upload
     to take place in batch.

- Simplifies the program API that the device must implement.

Note that for migration and compatibility purposes the old API
to set uniforms/samplers is still retained but completely on the
"client side" of the API and it just forwards the call to the
actual state object. When everything has been migrated this API
can be removed!