Refactor geometry API to be completely immutable
Refactor geometry API to be completely immutable. Change the way
the geometry data is generated so that the drawable objects will
now generate the data into a GeometryBuffer which is then used
as the data when creating the actual device specific geometry
object.

This means each geometry object is now immutable after created
and since we're providing them by a handle (shared_ptr) it will
now be possible to create these ahead of time, outside the rendering
loop and refer to them later on when rendering without the contents
getting messed up.