Add audio Queue element
Add audio Queue element. The queue can take copies of the buffers
coming in and hold them inside a queue. This allows a single source
to be split and queued for later use for some effect such as an echo.

TODO: To improve the performance the splitter could/should implement
COW semantics instead.