Add first version of audio graph preloading
Add first version of audio graph preloading. The idea is to decode
any FileSource with PCM caching enabled into PCM data so that it's
ready to use quickly without having to create a decoder. Preloading
functionality relies on the previously already implemented PCM
caching which means for this use case it could be improved with
more intellectual buffering, i.e. allocate a single buffer large
enough to hold the contents of the whole audio clip in PCM.