Enhance audio file source element with async decoder open
Enhance audio file source element with async decoder open. After
profiling it was discovered that the audio codec open can take
up to 4ms (profiled on the laptop) of walltime which will create
considerable problems in the tight main loop. We want to offload
this decoder open to a background thread. However, the problem is
that the audio graph prepare (element prepare) is a synchronous
operation and it must set the right format on the output ports.

Previous the details for setting the output format correctly
were only available after the decoder had been opened on the audio
stream.

This commit adds an audio graph preloading step which will open
every audio file stream and discover the information, sample rate,
channels and number of frames and store that in a cache. This
then allows us to use the cached information when preparing the
audio file source which means that the decoder open can be
referred to a background thread.

So basically:

- when we have cached information about the audio file
  create a background task for doing the actual decoder open and
  store the task handle for later.

- When audio data needs to be produced check whether we have a task
  for the audio decoder opening. If a task exists check whether it
  has completed (and w/o errors).

- When audio decoder open has completed, transfer the decoder object
  from the task to the file source element.

The only question right now is what to do if the decoder open hasn't
completed by the time Process is called. Some possible options:

- Do nothing (current solution)
- Generate silence (maybe this is the same as doing nothing?)
- Wait on the task handle?
  Probably a bad idea, because the Process function is called by
  the audio player thread which is "real time" thread.