Add Emscripten GameEngine thread pool size build parameter
Add Emscripten GameEngine thread pool size build parameter.

The thread pool (web worker) pool is important because the thread
creation must happen in the main browser thread. So if the browser
thread expexts to launch a thread and waits on it and never returns
to the browsers main event loop the thread will never start and the
code will hang indefinitely.

Care must be taken to match the parameter here to what is the
actual number of threads in the engine.

Currently we have the following threads:

- Audio player thread x 1
- Thread pool (background tasks) threads x 2