Refactor engine audio buffering strategy in the engine loader
Refactor engine audio buffering strategy in the engine loader.

There are a couple of problems here:

- We must make sure that the audio buffer (stream) is thread safe.

- On Emscripten the file system is GARBAGE and can only be accessed
  from the main browser (JS) thread. So any FS operation from another
  thread is silently proxied to the main thread.

  We can mitigate this issue by loading the audio buffer from the
  filesystem in the file pre-load stage and then keep the buffer
  around which will let the data/contents to be accessed without
  any filesystem access, thus letting the background threads to
  work better and not have to block/wait on the main thread on the
  FS operation.

- Remove the file mapping strategy on Emscripten. I'm sceptical
  this would work without stupid issues.

TODO: In the future it might make sense to create a custom
meta filesystem package that allows the game resources to be
packaged into a single file and then the resources would be accessed
from some offset in that "resource pack" that would be mapped or
loaded as a single buffer.