Refactor game engine audio threading strategy
Refactor game engine audio threading strategy. The problem is
that on Emscripten the backend audio implementation is OpenAL
which is implemented using Web Audio. The worker threads (web
workers) cannot use Web Audio but all access must happen
from the main thread.

This essentially means that on Emscripten having a player thread
access the device does not scale since the device access is silently
proxied to the main thread from the player thread and parallelism
is hampered.

Additionally there seems to be weird/obscure bugs that have appeared
after the audio threading was enabled. (Currently seen on Win10
with Firefox, Chrome and Edge). Either failure to load the web
worker JS script (browser complains about the mime type being
'text/plain' even though the resources are listed as 'text/javascript'.
The other bug was weird audio problems, glitches. The issues
seem related to the threading.

This commit changes the current audio threading strategy so that
the audio engine's audio streams (music and FX) are wrapped inside
threaded sources, and the audio player runs in the main thread.

This means that the sources *should* be able to run parallel and
produce PCM data without hickups and the device access (through
the player) will then run in the main thread and simply transfer
the PMC generated in the background threads to the device.