Add calls to flush loggers when a test case has failed
Add calls to flush loggers when a test case has failed and called
BlurpFailure and there are either thread local or global loggers.
This is trying to make sure that all the (hopefully useful) logging
information has actually been written to the file to support bug
analysis before we (possibly) terminate the process.

The extra complication here is that in order to avoid a build time
dependency to base/logging the log flush needs to be done through
dynamic function pointer resolution.