Add missing audio player ProcessOnce to the audio widget in editor
Add missing audio player ProcessOnce to the audio widget in editor.
If the audio player thread is *not* used we get no audio playback
unless the main thread then runs the audio device.

The problem with this "fix" here is that the audio is clippy since
the Qt main thread can be rather busy occasionally. Using the audio
thread in the editor though is currently tied to the same build
config as the game engine itself since the audio lib is only a
single build target.