Add background tasks to init and update particles in particle engine
Add background tasks to init and update particles in particle engine.

In order to avoid locking the "rendering" particle buffer for a
long time the background tasks copy the latest contents into a
temp buffer and then simply swap that with the rendering particle
buffer. This extra buffer turns a lock+copy into lock+swap.