Refactor uikit widget animation state management
Refactor uikit widget animation state management. Since this class
couples both the state that is the animation "design template" and
the actual execution state care needs to be taken in order to not
mess up the state.

Create new state arrays for the action states that get reset on
each animation trigger. This makes it easier to keep the running
state separate from the animation "design" state. It could be that
we should split this up too into "class" and "instance" but since
there are only ever instance per each class it's probably ok like
as-is for now anyways.

This commit also fixes 3 bugs that were present in the previous
implementation.

- An animation could be started too soon. For example if the user
  had several animation sequences under a single trigger where
  one would do changes and another sequence would undo the same
  changes, the first sequence could finish and then be restarted
  while the second sequence was still running thus leading to a
  messed up combined state.

  The fix right now is to not trigger any animations when there's
  even a single active animation on that widget under the same
  animation trigger.

- Loop counter state was not reset before re-trigger. This bug
  was caused by incorrect state management, i.e. messing up between
  the "class" state and the "instance/execution" state.

- Star values for interpolated animation properties were not properly
  re-initialized on re-trigger. On each animation re-trigger we
  must read the current start values in order to interpolate between
  the current "start" and the "end".

TODO: Since it's possible that the widget moves while the mouse doesn't
it's also possible that the widget is no longer under mouse after
some animation update. In fact the same problem could also happen
if someone called Move/SetPosition on the widget programmatically.

These conditions should trigger mouse event generation for correct
"under mouse" detection and MouseEnter/MouseLeave events.