Add a way to use the property system to define widget colors
Add a way to use the property system to define widget colors instead
of having to use the more complicated material interface. This is
only good for simple changes using a single color value, but the
caller interface is now much simpler.

When a material is looked up the widget property set is also then
checked for a property with the same key and "-color" suffix.