Add expected audio format to the audio device create
Add expected audio format to the audio device create. When using
OpenAL on Emscripten it seems that the JavaScript implementation
in library_openal.js doesn't use any 'samplerate' parameter when
creating the web AudioContext, which means that the default
samplerate will be used! Not setting the correct sample rate on
the audio context seems like a bug!

Use the OpenAL context parameters to set the samplerate with
ACL_FREQUENCY which seems to be parameter that controls the
audio context sample rate.