Refactor engine audio subsystem to use a single stream
Refactor engine audio subsystem to use a single stream. The reasons
behind this change are the following:

- Using a single stream only allows us to use an audio backend that
  only supports a single output stream. For example SDL or Sokol.

- If the audio backend is OpenAL on emscripten there's likely going
  to be less work to be done in the JS implementation since no
  mixing is required.

There are still some extra complications in this audio graph because
the gain control is a separate element so in order to control the
effects and music volume separately there still needs to be 2
separate streams.