Fix missing tracing stop in play window when exiting and tracing
Fix missing tracing stop in play window when exiting and tracing.
If the tracing is on the playwindow will delete the tracing objects
but leaves the global pointers with trash values. If the play
window runs in the editor process then bad things will happen on
the next game launch!