Add support for audio thread tracing in the engine and host apps
Add support for audio thread tracing in the engine and host apps.
This means that we have to manage the audio thread tracer state
including the thread specific tracer creation and deletion and
accessing the thread writer for dumping the output.

In order to be able to aggregate all tracing info into a single
the audio thread is given access to the same trace writer used by
the main thread except that the tracer is wrapped inside a locked
tracer in order to make it thread safe.