Fix possible dangling pointer when entering trace from Lua
Fix possible dangling pointer when entering trace from Lua, the
trace entry holds data via const char* and the pointer must remain
valid for the duration of the current trace iteration. When entering
any dynamic string (such as coming from Lua) we must make sure the
lifetime of the string matches the lifetime of the current trace
stack.

Add a new method in the tracer to store dynamic strings and return
a pointer for conveniently using in the trace entry. Clearing the
dynamic strings on each new trace stack iteration start.