Refactor entity node transformation and data away from entity
Refactor entity node transformation and data away from entity and into
separate structures that are then allocated throuh the entity node
allocator. The intention is to improve the cache locality when
accessing (transforming) entity nodes.

Currently each entity class holds the allocator instance for the
nodes of that entity type so that it's easy to associate transformation
functionality with a certain type of an entity. I.e. entity types
"Foo" and "Bar" can transform in different ways and that part of
the code doesn't need to check for type of the entity/node.