Add entity de-allocation/deletion background task
Add entity de-allocation/deletion background task. Since the spawning
can be async there can be contention on the allocator lock between
allocation and de-allocation. If the allocator is locked by a background
thread doing entity allocation then the main thread will be block
trying to de-allocate. So when the threadpool is in use create a
task to delete the entities and post it to some background thread
in order to avoid blocking the main thread.