Add first version of entity node transformer attachment
Add first version of entity node transformer attachment. The
indeded use case is for simple kinematic things such as an object
moving with constant velocity or with simple acceleration. Using a
node transformer allows one to express this motion and run it in
native code without having to rely on the dog slow Lua code for this.