Restructure the engine update and rendering loops
Restructure the engine update and rendering loops with the intention
to move things more forward a model where the "update" step can be
executed as a task in a separate thread and the time step used by
the update is then fed to the renderer in order to update the renderer
material animations etc.

The current blocker is that the renderer is still creating gfx
resources during the draw operation and accessess the entity/entity
node objects for this purpose. This would require locking and would
not produce the best parallelism. A refactoring should be done to
restructure the renderer so that no access to scene/entity/entity nodes
is needed during render.

Issue #195