Add a quick and dirty workspace loading progress dialog
Add a quick and dirty workspace loading progress dialog. Offload
the laoding to a std::thread that loads the content while the main
thread just pumps the Qt main loop inside a busy loop until the
loading thread is ready.

Add some #ifdef hacks for conditional building since we want this
only for the editor build and not in the unit tests.