Improve dependant Lua script packing using anchor from URI
Improve dependant Lua script packing using anchor from URI. If we
have a script that is a resource and has a URI such as
ws://lua/player/foobar.lua  which then refers to another script
with a relative file include such as require('foo/utility.lua')
try to look for the utility.lua relative to the foobar.lua
by creating an URI ws://lua/player/foo/utility.lua.

Note that this will still fail if we're packing a script that is
also in the same foo/ folder for example

ws://lua/player/foo/kek.lua which 'require('foo/utility.lua'
because then the new URI becomes ws://lua/player/foo/foo/utility.lua

This problem is currently unsolvable with the information we have.
Better to use proper URIs when require'ing the scripts in Lua.