Add read-only script var initialization support on entity create
Add read-only script var initialization support on entity create.
Temporarily turn off "read only" so that the variable can be initialized
in entity ctor. This is useful in combination with async spawn +
script vars that are passed in the spawn. Unfortunately this also
means that a copy of the variable then must be made in order to
allow the editing to take place.