Fix stale tile selection in editor tilemap widget
Fix stale tile selection in editor tilemap widget. If the map
layer changes it's possible that the selection coordinates are
no longer valid for the current layer if the layer's resolution
is different than the previous layer resolution.

Add checks on layer change to see if there's a selection and whether
the selection was for a layer with different resolution and if so
then clear the selection.