Fix incorrect FBO resolve target texture size after config change
Fix incorrect FBO resolve target texture size after config change.
When there's no client allocated texture assigned as the color
(resolve) target the FBO allocates one on the first call to Complete
the FBO and then assigns that texture pointer to the resolve target
texture pointer. However if the FBO is reconfigured and the size
changes then we must also check that if the resolve target is our
own resolve target from previous use then it must be re-allocated
if the FBO config dimensions have changed.