Add missing Boost package lookup in the Qt designer plugin build
Add missing Boost package lookup in the Qt designer plugin build.
This is due to some headers leaking some boost even though we should
not need it in the designer plugin (it should be possible to build
the plugin without these dependencies with some work). But for now
the easiest solution is just to use the boost package in the plugin
build since we need boost anyway for the engine/editor build.