Add some IPC hacks trying to make sure the game host syncs with editor
Add some IPC hacks trying to make sure the game host syncs with editor.

It seems that on some machine (testing on Win10) the last IPC messages
that the game host sends to sync the latest workspace properties
never arrive to the IPC host (to the editor).

Seems both sides have trouble:

- The client side exits the process too early without making sure
  the data is actually sent. Seems that Qt is borking and calling
  waitForBytesWritten(-1) doesn't work.
  => Add a shitty random loop to run the Qt event loop for some
     more time.

- The host side closes the IPC socket on process termination signal.
  => Add a timer delay to closing the IPC host socket trying to make
     sure that all data from the client has been read.
  - Can't use waitForBytesRead(-1) here because that could then
    block indefinitely so some more work would need to be done
    here.

Overall it seems that there are some Qt weirdness issues here too,
likely related to not having called exec() on the Qt app.

TODO: This probably needs some more work and a proper stateful
protocol would need to be implemented between the host and the
client.