Add some more logging and error checking in workspace property update
Add some more logging and error checking in workspace property update.
This update happens when an IPC message is received from the game
host after the game playback has ended in a separate process. Due
the shitty nature of the IPC there are various failure points here
and thus we need to make some more noise here when things are failing.