Add a hack to the playwindow log bits saving
Add a hack to the playwindow log bits saving. The user property value
turns into a qlonglong at some point (possibly when reading it back)
and the playwindow writes a quint. So this means that the property
values no longer actually match when the property is synced over IPC
back to the editor host from the game. This is a hack fix converting
from qint to qlong (qint64) trying to make the types match.